requirements.txt
Flask
Flask-SQLAlchemy
flask-restful
Flask-Migrate
Flask-Security-Too
Flask-JWT-Extended
Flask-Mail
Flask-Cors
Celery
Redis
SQLAlchemy
bcrypt
itsdangerous
PyJWT
